* More fixes...
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1704 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ creatureevent.cpp
@@ -710,7 +710,7 @@ uint32_t CreatureEvent::executeMailReceive(Player* player, Player* sender, Item*
 	}
 }
 
-uint32_t CreatureEvent::executeLook(Player* player, Thing* thing, const Position& position, uint8_t stackpos, int32_t lookDistance)
+uint32_t CreatureEvent::executeLook(Player* player, Thing* thing, const Position& position, int16_t stackpos, int32_t lookDistance)
 {
 	//onLook(cid, thing, position, lookDistance)
 	if(m_scriptInterface->reserveScriptEnv())
