+++ map.h
@@ -126,8 +126,8 @@ class QTreeLeafNode : public QTreeNode
 		QTreeLeafNode();
 		virtual ~QTreeLeafNode();
 
-		Floor* createFloor(uint8_t z);
-		Floor* getFloor(uint8_t z){return m_array[z];}
+		Floor* createFloor(uint16_t z);
+		Floor* getFloor(uint16_t z){return m_array[z];}
 
 		QTreeLeafNode* stepSouth(){return m_leafS;}
 		QTreeLeafNode* stepEast(){return m_leafE;}
@@ -179,14 +179,14 @@ class Map
 		* Get a single tile.
 		* \returns A pointer to that tile.
 		*/
-		Tile* getTile(uint16_t x, uint16_t y, uint8_t z);
+		Tile* getTile(uint16_t x, uint16_t y, uint16_t z);
 		Tile* getTile(const Position& pos);
 
 		/**
 		* Set a single tile.
 		* \param a tile to set for the position
 		*/
-		void setTile(uint16_t _x, uint16_t _y, uint8_t _z, Tile* newTile);
+		void setTile(uint16_t _x, uint16_t _y, uint16_t _z, Tile* newTile);
 		void setTile(const Position& pos, Tile* newTile)
 		{
 			setTile(pos.x, pos.y, pos.z, newTile);
